Implement the salesmen handler
