+++ backend-ts/src/bll/index.ts
@@ -8,4 +8,5 @@
 export * from "./errors.js"
 export * from "./products.js"
 export * from "./rules.js"
+export * from "./salesmen.js"
 export * from "./validator.js"
